+++ ChangeLog.md
@@ -10,7 +10,9 @@
 * Added {Ore::Template::Helpers#gem_package_task?}.
 * Added {Ore::Template::Directory#dependencies}.
 * Added {Ore::Template::Directory#development_dependencies}.
+* Added {Ore::Generator::OPTIONS}.
 * Renamed `Ore::Config.default_options` to {Ore::Config.options}.
+* Renamed `Ore::Generator.defaults` to {Ore::Generator.options}.
 * Renamed `Ore::Generator.templates` to {Ore::Template.templates}.
 * Renamed `Ore::Generator.template?` to {Ore::Template.template?}.
 * Renamed `Ore::Generator.register_template` to {Ore::Template.register}.
@@ -119,7 +121,7 @@
 * Added {Ore::Config.enable!}.
 * Added {Ore::Config.disable!}.
 * Added `Ore::Config.default_options`.
-* Added {Ore::Generator.defaults}.
+* Added `Ore::Generator.defaults`.
 * Added {Ore::Generator.generator_option}.
 * Added `vendor/cache/*.gem` to `.gitignore` if `--bundler` is specified.
 * Attempt to auto-load `ore/specification` in the generated `*.gemspec`